requirements.txt
faster-whisper
rapidocr-onnxruntime
yt-dlp
python-docx
